Use custom property field 'publishing' instead of the existing 'status' field for document
The individual document repositories are changing the status within their respective `push()` function.
This conflicts with the desired behavior of having the 'publishing' status operate on a purely document level.
To avoid any major redesigns, just implement a custom boolean field (database column) for designating the state for when content is currently publishing.
This change requires front-end changes.

Update the java tests appropriately.