Item 03983: Broadcast last event to in-progress Publishing box - eliminate the spinner
Create publishing repository specific events.
Broadcast these events for the Dspace repository.

The broadcasting channel is implemented at "/channel/publishing" because "/channel/repository" seems a bit confusing due to there being conflicting meanings of "repo" in use.
The document channel is then at "/channel/publishing/document" to avoid id collisions for anything else that potentially may need use a publishing channel.

A PublishingType enum is implemented to give hints to the front-end on how to process and present specific events.
This PublishingType enum is implemented very generally to allow being used in other repositories, such as Fedora.

The Dspace Repository events are broadcast with detailed information.
The Repository being connected/disconnected to is specifically added to help ensure that when multiple repositories are used, the log will be less confusing.

There are some cases where the session is not closed, namely on Exceptions.
Ensure that the connection is closed and present the event log for the end of the session.

The isPublishing state should always be reset on exception.